Move writer interface to fmt package.
Added a simple json writer for possible debugging.
